PEP 396 compatible __version__ attribute.
PEP396: Module Version Numbers tells us that module version numbers should be available in a __version__ attribute.  Historically, pySerial has used VERSION instead. This commit adds __version__ and aliases the legacy VERSION to it for backwards compatibility.

Note that although PEP396 is not yet accepted, this is a very widely followed standard and I was surprised that pySerial did not follow it.