Avoids setup.py having to import the serial package.
It's bad practice for setup.py to import the package it is
installing in order to get, say, version information. For
example. if PYTHONPATH has been manipulated, the
import serial statement may not in fact import the
expected version.

This commit uses a technique for single-sourcing the
version number as recommended in the Python Packaging
Guide at

    http://python-packaging-user-guide.readthedocs.io/en/latest/single_source_version/

In essence, we read serial.__init__.py as a text file and
parse it for the version number, rather than importing it.