Only install serial.aio on Python >= 3.4.
Changes serial/aio.py into a package serial/aio/__init__.py.
This allows us to use distutils or setuptools to optionally install this package.  The changes to setup.py detect whether the current Python version is >= 3.5 and if so append serial.aio to the list of packages to be installed.

Be aware that when testing this you MUST has previously cleaned any existing distutils/setuptools build directory with:

    python setup.py clean --all

 otherwise the old aio may continue to be installed.