miniterm.py:  Adds execute permissions.
This has a suitable shebang line to be run directly as a program, but
needs execute permissions to make that actually work.

The same change could be made to list_ports*, but it's less important
there where command-line invocation is only for test purposes.

TESTED:
./miniterm.py now launches on macOS.  Other POSIX systems should
behave the same.  No idea what effect this has on Windows.