refactor checks for valid timeouts
This unifies the error paths for the two checks done on new timeout
values. The resulting checking is slightly stricter, but I don't expect
a real impact in practice. (For example complex numbers and timeouts of
type decimal.Decimal are refused now with the former resulting in
problems afterwards.)

This was suggested by penguinolog in a review comment to
https://github.com/pyserial/pyserial/pull/382.