Changes to "IOUSBInterface" and "USB Interface Name" on Apple Silicon
On M1 the interface object is now reported by `IOObjectGetClass()` as a `IOUSBHostInterface`.
And the `USB Interface Name` property is gone. But there is a `kUSBString` property with the same content.

Also cleanup to remove `search_for_locationID_in_interfaces` since I'm more confident it is rendered useless now (and would need to be fixed for M1 anyway).