fixed an issue with loading symlinked devices on Linux
~ serial.tools.list_ports.comports(include_links)
On Linux, even when `include_links=False`, links are still included in
the returned list, just like when `include_links=True`.
The problem is that `serial.tools.list_ports_linux.comports` does not
expect any of the files loaded from its predefined paths to be a link
(which is a false assumption).

The solution to this seems to be obvious: we *consciously* skip all
links in the first load, leaving all link loading to
`serial/tools/list_ports_common.list_links`.