Close opened file handler on closing the serial port
When using the  spy://___URI___?file=__SOME_FILE___&raw format for serial port the __SOME_FILE___ is opened and its file handler is not closed when the serial port is closed. This patch fixes this. By default the opened stream is 'stderr' which cannot be closed, so check for its existence is made.