Override RawIOBase.close() and .closed
Inheriting from RawIOBase means we inherit its finalization behavior,
whereby close() is called if .closed is False. Because Serial does not
implement .closed, we get the RawIOBase implementation which does not
reflect the fact that Serial can be opened/closed multiple times.

We now provide a .closed property which reflects .is_open. This way, if
Serial thinks it is closed, RawIOBase will not call close() again.

This also implements close() to override RawIOBase.closed, as RawIOBase
can only be closed once.

Fixes #670